sort products alphabetically on the recipe edit page

+++ controllers/bo/component/ecommerce/recipe_ingredients.php
@@ -30,17 +30,28 @@ class Onxshop_Controller_Bo_Component_Ecommerce_Recipe_Ingredients extends Onxsh
 		if (is_array($_POST['ingredients'])) {
 
 			$current = $Ingredients->listing("recipe_id = $recipe_id");
+			$keep = array_keys($_POST['ingredients']);
 
 			foreach ($current as $c) {
-				$Ingredients->delete($c['id']);
+				if (!in_array($c['id'], $keep)) $Ingredients->delete($c['id']);
 			}
 		
-			foreach ($_POST['ingredients'] as $product_id => $item) {
-				if (is_numeric($product_id)) {
-					$ingredients['product_id'] = $product_id;
+			foreach ($_POST['ingredients'] as $ingredient_id => $item) {
+				if (is_numeric($ingredient_id)) {
+					$ingredients['id'] = $ingredient_id;
+					$ingredients['product_variety_id'] = $item['product_variety_id'];
 					$ingredients['quantity'] = $item['quantity'];
 					$ingredients['units'] = $item['units'];
 					$ingredients['notes'] = $item['notes'];
+					$ingredients['group_title'] = $item['group_title'];
+					$Ingredients->update($ingredients);
+				} else {
+					unset($ingredients['id']);
+					$ingredients['product_variety_id'] = $item['product_variety_id'];
+					$ingredients['quantity'] = $item['quantity'];
+					$ingredients['units'] = $item['units'];
+					$ingredients['notes'] = $item['notes'];
+					$ingredients['group_title'] = $item['group_title'];
 					$Ingredients->insert($ingredients);
 				}
 			}
@@ -52,16 +63,39 @@ class Onxshop_Controller_Bo_Component_Ecommerce_Recipe_Ingredients extends Onxsh
 		$units = $Ingredients->getUnits();
 		$this->parseUnits($units, false, 'head.unit');
 
+
+		/**
+		 * get ingredient list (products)
+		 */
+		
+		$products = $Product->getProductList();
+		$products = php_multisort($products, array(array("key" => "name", "sort" => "asc")));
+		$this->parseIngredients($products, false, 'head.product');
+
+		/**
+		 * render template
+		 */
+		
+		foreach ($products as $product) {
+			$this->tpl->assign("PRODUCT", $product);
+			foreach ($product['variety'] as $variety) {
+				$variety['selected'] = $variety['id'] == $ingredient['product_variety_id'] ? 'selected="selected"' : '';
+				$this->tpl->assign("VARIETY", $variety);
+				$this->tpl->parse("content.template.product");
+			}
+		}
+		$this->tpl->parse("content.template");
+
 		/**
 		 * listing
 		 */
 		 
 		$current = $Ingredients->listing("recipe_id = $recipe_id");
+
 		foreach ($current as $ingredient) {
-			$detail = $Product->detail($ingredient['product_id']);
 			if ($detail['publish'] == 0) $detail['class'] = "class='disabled'";
-			$this->tpl->assign("PRODUCT", $detail);
 			$this->tpl->assign("ITEM", $ingredient);
+			$this->parseIngredients($products, $ingredient['product_variety_id']);
 			$this->parseUnits($units, $ingredient['units']);
 			$this->tpl->parse("content.item");
 		}
@@ -76,5 +110,20 @@ class Onxshop_Controller_Bo_Component_Ecommerce_Recipe_Ingredients extends Onxsh
 			$this->tpl->parse($block);
 		}
 	}
+
+	public function parseIngredients(&$products, $active, $block = 'content.item.product') {
+
+		foreach ($products as $product) {
+			$this->tpl->assign("PRODUCT", $product);
+			foreach ($product['variety'] as $variety) {
+				if ($product['publish'] == 1 && $variety['publish'] == 1) {
+					$variety['selected'] = $variety['id'] == $active ? 'selected="selected"' : '';
+					$this->tpl->assign("VARIETY", $variety);
+					$this->tpl->parse($block);
+				}
+			}
+		}
+	}
+
 }
